+++ oci/core_instance_resource.go
@@ -133,6 +133,11 @@ func CoreInstanceResource() *schema.Resource {
 					},
 				},
 			},
+			"capacity_reservation_id": {
+				Type:     schema.TypeString,
+				Optional: true,
+				Computed: true,
+			},
 			"create_vnic_details": {
 				Type:     schema.TypeList,
 				Optional: true,
@@ -731,6 +736,11 @@ func (s *CoreInstanceResourceCrud) Create() error {
 		request.AvailabilityDomain = &tmp
 	}
 
+	if capacityReservationId, ok := s.D.GetOkExists("capacity_reservation_id"); ok {
+		tmp := capacityReservationId.(string)
+		request.CapacityReservationId = &tmp
+	}
+
 	if compartmentId, ok := s.D.GetOkExists("compartment_id"); ok {
 		tmp := compartmentId.(string)
 		request.CompartmentId = &tmp
@@ -964,6 +974,11 @@ func (s *CoreInstanceResourceCrud) Update() error {
 		}
 	}
 
+	if capacityReservationId, ok := s.D.GetOkExists("capacity_reservation_id"); ok {
+		tmp := capacityReservationId.(string)
+		request.CapacityReservationId = &tmp
+	}
+
 	if definedTags, ok := s.D.GetOkExists("defined_tags"); ok {
 		convertedDefinedTags, err := mapToDefinedTags(definedTags.(map[string]interface{}))
 		if err != nil {
@@ -1110,6 +1125,10 @@ func (s *CoreInstanceResourceCrud) SetData() error {
 
 	if s.Res.AvailabilityDomain != nil {
 		s.D.Set("availability_domain", *s.Res.AvailabilityDomain)
+	}
+
+	if s.Res.CapacityReservationId != nil {
+		s.D.Set("capacity_reservation_id", *s.Res.CapacityReservationId)
 	}
 
 	if s.Res.CompartmentId != nil {